test: split form-controller tests by tool and extract shared fixtures
