+++ vaadin-ai-components-flow-parent/vaadin-ai-components-flow/src/test/java/com/vaadin/flow/component/ai/form/FormAIControllerTest.java
@@ -15,18 +15,12 @@
  */
 package com.vaadin.flow.component.ai.form;
 
-import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.util.ArrayList;
-import java.util.Arrays;
+import static com.vaadin.flow.component.ai.form.FormTestSupport.executeQueryFieldOptions;
+import static com.vaadin.flow.component.ai.form.FormTestSupport.findTool;
+import static com.vaadin.flow.component.ai.form.FormTestSupport.json;
+
 import java.util.Collection;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.BiFunction;
 import java.util.function.Function;
 
@@ -34,636 +28,21 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-import com.vaadin.flow.component.AbstractField;
-import com.vaadin.flow.component.Component;
-import com.vaadin.flow.component.ComponentUtil;
-import com.vaadin.flow.component.HasComponents;
-import com.vaadin.flow.component.HasHelper;
-import com.vaadin.flow.component.HasLabel;
-import com.vaadin.flow.component.HasValue;
-import com.vaadin.flow.component.ItemLabelGenerator;
-import com.vaadin.flow.component.Tag;
-import com.vaadin.flow.component.ai.provider.LLMProvider;
+import com.vaadin.flow.component.ai.form.FormTestFields.CompositeField;
+import com.vaadin.flow.component.ai.form.FormTestFields.IntField;
+import com.vaadin.flow.component.ai.form.FormTestFields.SingleSelectField;
+import com.vaadin.flow.component.ai.form.FormTestFields.TestField;
 import com.vaadin.flow.component.html.Div;
-import com.vaadin.flow.component.textfield.EmailField;
-import com.vaadin.flow.component.textfield.PasswordField;
-import com.vaadin.flow.data.binder.HasItems;
-import com.vaadin.flow.data.provider.CallbackDataProvider;
-import com.vaadin.flow.data.provider.DataProvider;
-import com.vaadin.flow.data.provider.HasDataView;
-import com.vaadin.flow.data.provider.HasLazyDataView;
-import com.vaadin.flow.data.provider.HasListDataView;
-import com.vaadin.flow.data.provider.ListDataProvider;
-import com.vaadin.flow.data.selection.MultiSelect;
-import com.vaadin.flow.data.selection.MultiSelectionListener;
 import com.vaadin.flow.internal.JacksonUtils;
-import com.vaadin.flow.shared.Registration;
 
-import tools.jackson.databind.JsonNode;
-
+/**
+ * Tests covering {@link FormAIController}'s construction, container traversal,
+ * field-locking lifecycle, and hint-registration API. Tool-output specifics for
+ * {@code get_form_state} and {@code query_field_options} live in
+ * {@link FormStateToolTest} and {@link QueryFieldOptionsToolTest} so this class
+ * stays focused on controller behaviour rather than schema details.
+ */
 class FormAIControllerTest {
-
-    /**
-     * Minimal {@link com.vaadin.flow.component.HasValue} component used by the
-     * discovery tests. {@link AbstractField} ships with flow-server, so
-     * exercising discovery does not require any concrete Vaadin input component
-     * module.
-     */
-    @Tag("test-field")
-    private static class TestField extends AbstractField<TestField, String> {
-        TestField() {
-            super("");
-        }
-
-        @Override
-        protected void setPresentationValue(String value) {
-        }
-    }
-
-    /**
-     * Composite field that is both {@link com.vaadin.flow.component.HasValue}
-     * and {@link HasComponents} — used to verify that discovery stops at the
-     * field and does not descend into the field's internal composition.
-     */
-    @Tag("composite-field")
-    private static class CompositeField extends
-            AbstractField<CompositeField, String> implements HasComponents {
-        CompositeField(Component... children) {
-            super("");
-            add(children);
-        }
-
-        @Override
-        protected void setPresentationValue(String value) {
-        }
-    }
-
-    /**
-     * Integer-valued field used to exercise {@code valueOptions} with a
-     * non-{@link String} value type, where {@code toValue} must convert the
-     * chosen label into the field's actual value type.
-     */
-    @Tag("int-field")
-    private static class IntField extends AbstractField<IntField, Integer> {
-        IntField() {
-            super(0);
-        }
-
-        @Override
-        protected void setPresentationValue(Integer value) {
-        }
-    }
-
-    @Tag("double-field")
-    private static class DoubleField
-            extends AbstractField<DoubleField, Double> {
-        DoubleField() {
-            super(null);
-        }
-
-        @Override
-        protected void setPresentationValue(Double value) {
-        }
-    }
-
-    @Tag("bigint-field")
-    private static class BigIntField
-            extends AbstractField<BigIntField, java.math.BigInteger> {
-        BigIntField() {
-            super(null);
-        }
-
-        @Override
-        protected void setPresentationValue(java.math.BigInteger value) {
-        }
-    }
-
-    @Tag("bigdec-field")
-    private static class BigDecField
-            extends AbstractField<BigDecField, BigDecimal> {
-        BigDecField() {
-            super(null);
-        }
-
-        @Override
-        protected void setPresentationValue(BigDecimal value) {
-        }
-    }
-
-    @Tag("bool-field")
-    private static class BoolField extends AbstractField<BoolField, Boolean> {
-        BoolField() {
-            super(false);
-        }
-
-        @Override
-        protected void setPresentationValue(Boolean value) {
-        }
-    }
-
-    @Tag("date-field")
-    private static class DateField extends AbstractField<DateField, LocalDate> {
-        DateField() {
-            super(null);
-        }
-
-        @Override
-        protected void setPresentationValue(LocalDate value) {
-        }
-    }
-
-    @Tag("datetime-field")
-    private static class DateTimeField
-            extends AbstractField<DateTimeField, LocalDateTime> {
-        DateTimeField() {
-            super(null);
-        }
-
-        @Override
-        protected void setPresentationValue(LocalDateTime value) {
-        }
-    }
-
-    @Tag("time-field")
-    private static class TimeField extends AbstractField<TimeField, LocalTime> {
-        TimeField() {
-            super(null);
-        }
-
-        @Override
-        protected void setPresentationValue(LocalTime value) {
-        }
-    }
-
-    /**
-     * Domain-typed item record used to exercise label-generator-driven
-     * rendering of selection options. Non-String item types are what makes the
-     * {@code ItemLabelGenerator} path observable in tests.
-     */
-    private record Project(String code, String name) {
-    }
-
-    @Tag("long-field")
-    private static class LongField extends AbstractField<LongField, Long> {
-        LongField() {
-            super(null);
-        }
-
-        @Override
-        protected void setPresentationValue(Long value) {
-        }
-    }
-
-    @Tag("short-field")
-    private static class ShortField extends AbstractField<ShortField, Short> {
-        ShortField() {
-            super(null);
-        }
-
-        @Override
-        protected void setPresentationValue(Short value) {
-        }
-    }
-
-    @Tag("byte-field")
-    private static class ByteField extends AbstractField<ByteField, Byte> {
-        ByteField() {
-            super(null);
-        }
-
-        @Override
-        protected void setPresentationValue(Byte value) {
-        }
-    }
-
-    @Tag("float-field")
-    private static class FloatField extends AbstractField<FloatField, Float> {
-        FloatField() {
-            super(null);
-        }
-
-        @Override
-        protected void setPresentationValue(Float value) {
-        }
-    }
-
-    /**
-     * Non-generic interface that pre-binds {@link HasValue}'s value type to
-     * {@link Integer}. Used to exercise the {@code findHasValueValueArg} walk
-     * when an intermediate interface in the hierarchy is not parameterized: the
-     * recursion must propagate the resolved type out of the raw-class branch
-     * instead of dropping it.
-     */
-    private interface IntegerSelectableField
-            extends HasValue<HasValue.ValueChangeEvent<Integer>, Integer> {
-    }
-
-    /**
-     * Field that implements {@link IntegerSelectableField} (a raw class in the
-     * {@code getGenericInterfaces()} view of this class) instead of
-     * {@code HasValue<?, Integer>} directly. Without the recursive
-     * {@code return found;} in {@code findHasValueValueArg}, the walk would
-     * find {@code Integer} but discard it, leaving the field classified as
-     * {@link FormFieldType#STRING}.
-     */
-    @Tag("integer-via-non-generic")
-    private static class IntegerViaNonGenericInterfaceField extends Component
-            implements IntegerSelectableField {
-
-        private Integer value;
-
-        @Override
-        public void setValue(Integer value) {
-            this.value = value;
-        }
-
-        @Override
-        public Integer getValue() {
-            return value;
-        }
-
-        @Override
-        public Registration addValueChangeListener(
-                HasValue.ValueChangeListener<? super HasValue.ValueChangeEvent<Integer>> listener) {
-            return () -> {
-            };
-        }
-
-        @Override
-        public boolean isReadOnly() {
-            return false;
-        }
-
-        @Override
-        public void setReadOnly(boolean readOnly) {
-        }
-
-        @Override
-        public boolean isRequiredIndicatorVisible() {
-            return false;
-        }
-
-        @Override
-        public void setRequiredIndicatorVisible(boolean visible) {
-        }
-    }
-
-    /**
-     * Minimal field that implements {@link HasLazyDataView} but
-     * <strong>not</strong> {@link HasItems}. Real Vaadin selection components
-     * (e.g. {@code ComboBox}) follow this shape — they expose the data-view
-     * marker interfaces but not {@code HasItems}. The classifier must recognise
-     * such fields as {@link FormFieldType#SINGLE_SELECT} via the data-view
-     * markers; falling back to {@link HasItems} alone would misclassify them as
-     * {@link FormFieldType#STRING}, suppressing the {@code enum} block built
-     * from the backing {@link ListDataProvider}.
-     */
-    @Tag("lazy-data-view-field")
-    private static class LazyDataViewField
-            extends AbstractField<LazyDataViewField, String> implements
-            HasLazyDataView<String, String, com.vaadin.flow.data.provider.LazyDataView<String>> {
-
-        private ListDataProvider<String> provider = DataProvider
-                .ofCollection(List.of());
-
-        LazyDataViewField() {
-            super("");
-        }
-
-        @Override
-        protected void setPresentationValue(String value) {
-        }
-
-        @Override
-        public com.vaadin.flow.data.provider.LazyDataView<String> setItems(
-                com.vaadin.flow.data.provider.BackEndDataProvider<String, String> dataProvider) {
-            return null;
-        }
-
-        @Override
-        public com.vaadin.flow.data.provider.LazyDataView<String> getLazyDataView() {
-            return null;
-        }
-
-        void supplyItems(String... items) {
-            provider = DataProvider.ofCollection(Arrays.asList(items));
-        }
-
-        @SuppressWarnings("unused")
-        public DataProvider<String, ?> getDataProvider() {
-            return provider;
-        }
-    }
-
-    /**
-     * Minimal field that implements only {@link HasListDataView}. Pinning each
-     * data-view marker separately catches regressions where one interface is
-     * dropped from the classifier's marker list while the others stay.
-     */
-    @Tag("list-data-view-field")
-    private static class ListDataViewField
-            extends AbstractField<ListDataViewField, String> implements
-            HasListDataView<String, com.vaadin.flow.data.provider.ListDataView<String, ?>> {
-
-        private ListDataProvider<String> provider = DataProvider
-                .ofCollection(List.of());
-
-        ListDataViewField() {
-            super("");
-        }
-
-        @Override
-        protected void setPresentationValue(String value) {
-        }
-
-        @Override
-        public com.vaadin.flow.data.provider.ListDataView<String, ?> setItems(
-                ListDataProvider<String> dataProvider) {
-            return null;
-        }
-
-        @Override
-        public com.vaadin.flow.data.provider.ListDataView<String, ?> getListDataView() {
-            return null;
-        }
-
-        void supplyItems(String... items) {
-            provider = DataProvider.ofCollection(Arrays.asList(items));
-        }
-
-        @SuppressWarnings("unused")
-        public DataProvider<String, ?> getDataProvider() {
-            return provider;
-        }
-    }
-
-    /**
-     * Minimal field that implements only {@link HasDataView}. See
-     * {@link LazyDataViewField} for rationale.
-     */
-    @Tag("data-view-field")
-    private static class DataViewField
-            extends AbstractField<DataViewField, String> implements
-            HasDataView<String, String, com.vaadin.flow.data.provider.DataView<String>> {
-
-        private ListDataProvider<String> provider = DataProvider
-                .ofCollection(List.of());
-
-        DataViewField() {
-            super("");
-        }
-
-        @Override
-        protected void setPresentationValue(String value) {
-        }
-
-        @Override
-        public com.vaadin.flow.data.provider.DataView<String> setItems(
-                DataProvider<String, String> dataProvider) {
-            return null;
-        }
-
-        @Override
-        public com.vaadin.flow.data.provider.DataView<String> setItems(
-                com.vaadin.flow.data.provider.InMemoryDataProvider<String> dataProvider) {
-            return null;
-        }
-
-        @Override
-        public com.vaadin.flow.data.provider.DataView<String> getGenericDataView() {
-            return null;
-        }
-
-        void supplyItems(String... items) {
-            provider = DataProvider.ofCollection(Arrays.asList(items));
-        }
-
-        @SuppressWarnings("unused")
-        public DataProvider<String, ?> getDataProvider() {
-            return provider;
-        }
-    }
-
-    /**
-     * String-valued field that also implements {@link HasLabel} and
-     * {@link HasHelper}, used by the description-merging test.
-     */
-    @Tag("labeled-string-field")
-    private static class LabeledStringField
-            extends AbstractField<LabeledStringField, String>
-            implements HasLabel, HasHelper {
-        LabeledStringField() {
-            super("");
-        }
-
-        @Override
-        protected void setPresentationValue(String value) {
-        }
-    }
-
-    @Tag("throwing-field")
-    private static class ThrowingField
-            extends AbstractField<ThrowingField, String> {
-        ThrowingField() {
-            super("");
-        }
-
-        @Override
-        public String getValue() {
-            throw new RuntimeException("boom");
-        }
-
-        @Override
-        protected void setPresentationValue(String value) {
-        }
-    }
-
-    /**
-     * A {@link Component} that implements {@link HasValue} directly — not via
-     * {@link AbstractField}. Used to verify that
-     * {@link FormFieldType#classify(HasValue)} resolves the value type when the
-     * {@code HasValue<?, V>} parameterization sits at the class declaration
-     * itself rather than several layers up the hierarchy.
-     */
-    @Tag("direct-integer-field")
-    private static class DirectIntegerField extends Component
-            implements HasValue<DirectIntegerField.ChangeEvent, Integer> {
-
-        private Integer value;
-
-        @Override
-        public void setValue(Integer value) {
-            this.value = value;
-        }
-
-        @Override
-        public Integer getValue() {
-            return value;
-        }
-
-        @Override
-        public Registration addValueChangeListener(
-                HasValue.ValueChangeListener<? super ChangeEvent> listener) {
-            return () -> {
-            };
-        }
-
-        @Override
-        public boolean isReadOnly() {
-            return false;
-        }
-
-        @Override
-        public void setReadOnly(boolean readOnly) {
-        }
-
-        @Override
-        public boolean isRequiredIndicatorVisible() {
-            return false;
-        }
-
-        @Override
-        public void setRequiredIndicatorVisible(boolean visible) {
-        }
-
-        static class ChangeEvent implements HasValue.ValueChangeEvent<Integer> {
-            @Override
-            public HasValue<?, Integer> getHasValue() {
-                return null;
-            }
-
-            @Override
-            public boolean isFromClient() {
-                return false;
-            }
-
-            @Override
-            public Integer getOldValue() {
-                return null;
-            }
-
-            @Override
-            public Integer getValue() {
-                return null;
-            }
-        }
-    }
-
-    /**
-     * Single-select field. Implements {@link HasItems} so the controller
-     * classifies it as a selection component, and exposes
-     * {@code getDataProvider()}/{@code getItemLabelGenerator()} reflectively so
-     * {@link FormValueConverter} can read items and labels.
-     */
-    @Tag("single-select-field")
-    private static class SingleSelectField<T> extends
-            AbstractField<SingleSelectField<T>, T> implements HasItems<T> {
-
-        private DataProvider<T, ?> provider = DataProvider
-                .ofCollection(List.of());
-        private ItemLabelGenerator<T> labelGenerator;
-
-        SingleSelectField() {
-            super(null);
-        }
-
-        @Override
-        protected void setPresentationValue(T value) {
-        }
-
-        @Override
-        public void setItems(Collection<T> items) {
-            provider = DataProvider.ofCollection(items);
-        }
-
-        @SafeVarargs
-        @SuppressWarnings("varargs")
-        @Override
-        public final void setItems(T... items) {
-            setItems(Arrays.asList(items));
-        }
-
-        void setDataProvider(DataProvider<T, ?> provider) {
-            this.provider = provider;
-        }
-
-        @SuppressWarnings("unused")
-        public DataProvider<T, ?> getDataProvider() {
-            return provider;
-        }
-
-        @SuppressWarnings("unused")
-        public ItemLabelGenerator<T> getItemLabelGenerator() {
-            return labelGenerator;
-        }
-
-        public void setItemLabelGenerator(ItemLabelGenerator<T> generator) {
-            this.labelGenerator = generator;
-        }
-    }
-
-    /**
-     * Multi-select field. Implements {@link MultiSelect} so the controller
-     * classifies it as a multi-select, and exposes the same reflective
-     * accessors as {@link SingleSelectField} for items and label generation.
-     */
-    @Tag("multi-select-field")
-    private static class MultiSelectField<T>
-            extends AbstractField<MultiSelectField<T>, Set<T>>
-            implements MultiSelect<MultiSelectField<T>, T> {
-
-        private ListDataProvider<T> provider = DataProvider
-                .ofCollection(List.of());
-        private ItemLabelGenerator<T> labelGenerator;
-
-        MultiSelectField() {
-            super(Set.of());
-        }
-
-        @Override
-        protected void setPresentationValue(Set<T> value) {
-        }
-
-        @SafeVarargs
-        final void setItems(T... items) {
-            provider = DataProvider.ofCollection(Arrays.asList(items));
-        }
-
-        @Override
-        public void updateSelection(Set<T> added, Set<T> removed) {
-            var next = new HashSet<>(getValue());
-            next.addAll(added);
-            next.removeAll(removed);
-            setValue(next);
-        }
-
-        @Override
-        public Set<T> getSelectedItems() {
-            return getValue();
-        }
-
-        @Override
-        public Registration addSelectionListener(
-                MultiSelectionListener<MultiSelectField<T>, T> listener) {
-            return () -> {
-            };
-        }
-
-        @SuppressWarnings("unused")
-        public ListDataProvider<T> getDataProvider() {
-            return provider;
-        }
-
-        @SuppressWarnings("unused")
-        public ItemLabelGenerator<T> getItemLabelGenerator() {
-            return labelGenerator;
-        }
-
-        @SuppressWarnings("unused")
-        public void setItemLabelGenerator(ItemLabelGenerator<T> generator) {
-            this.labelGenerator = generator;
-        }
-    }
 
     @Nested
     class Construction {
@@ -886,8 +265,8 @@ class FormAIControllerTest {
 
             Assertions.assertTrue(field.isReadOnly(),
                     "A field the application set read-only between turns "
-                            + "must stay read-only after a subsequent "
-                            + "fill releases its own locks");
+                            + "must stay read-only after a subsequent fill "
+                            + "releases its own locks");
         }
 
         @Test
@@ -906,18 +285,18 @@ class FormAIControllerTest {
 
             Assertions.assertTrue(initial.isReadOnly());
             Assertions.assertTrue(added.isReadOnly(),
-                    "Fields added between turns must be locked on the "
-                            + "next request");
+                    "Fields added between turns must be locked on the next "
+                            + "request");
         }
 
         @Test
         void fieldIgnoredBetweenTurnsIsNotLockedOnNextRequest() {
             // The application may flag a field as ignored after the
-            // controller has been wired up — e.g., a feature toggle
-            // hides PII from the AI. The next turn must respect that.
-            // Today this works only because discovery re-evaluates each
-            // request; if anyone caches the active set "for performance",
-            // this regression slips through.
+            // controller has been wired up — e.g., a feature toggle hides
+            // PII from the AI. The next turn must respect that. Today this
+            // works only because discovery re-evaluates each request; if
+            // anyone caches the active set "for performance", this
+            // regression slips through.
             var field = new TestField();
             var controller = new FormAIController(new Div(field));
 
@@ -963,15 +342,14 @@ class FormAIControllerTest {
 
             Assertions.assertEquals("banana\n",
                     executeQueryFieldOptions(controller, field, "an", 10),
-                    "Filter must restrict results to options containing "
-                            + "the filter substring");
+                    "Filter must restrict results to options containing the "
+                            + "filter substring");
             Assertions.assertEquals("banana\n",
                     executeQueryFieldOptions(controller, field, "AN", 10),
                     "Filter must match labels regardless of filter case");
             Assertions.assertEquals("apple\nbanana\ncherry\n",
                     executeQueryFieldOptions(controller, field, "", 10),
-                    "Empty filter must return all options up to the "
-                            + "limit");
+                    "Empty filter must return all options up to the limit");
         }
 
         @Test
@@ -1035,10 +413,11 @@ class FormAIControllerTest {
         @Test
         void reregisteringWithBiFunctionClearsPriorFixedOptionsFlag() {
             // Each valueOptions call replaces the previous registration for
-            // the same field. The fixed-collection overload sets a flag that
-            // makes the schema render options as 'enum'; re-registering with
-            // a BiFunction must reset that flag so the schema rendering
-            // matches the new registration (queryable, not enum).
+            // the same field. The fixed-collection overload sets a flag
+            // that makes the schema render options as 'enum'; re-
+            // registering with a BiFunction must reset that flag so the
+            // schema rendering matches the new registration (queryable, not
+            // enum).
             var combo = new SingleSelectField<String>();
             var controller = new FormAIController(new Div(combo));
             controller.valueOptions(combo, List.of("EUR", "USD"));
@@ -1056,1345 +435,5 @@ class FormAIControllerTest {
                     "Stale enum block must not survive re-registration with "
                             + "a BiFunction, got: " + field);
         }
-    }
-
-    @Nested
-    class GetTools {
-
-        @Test
-        void queryFieldOptionsReturnsRegisteredOptions() {
-            // End-to-end: register valueOptions on a field, then drive the
-            // tool the way an LLM would — call getTools().execute(...) with
-            // the field id. Pins the wiring from valueOptions registration
-            // through ToolCallbacks to the query function.
-            var field = new TestField();
-            var controller = new FormAIController(new Div(field));
-            controller.valueOptions(field, List.of("apple", "banana", "cherry"),
-                    Function.identity());
-            controller.onRequestStart();
-
-            var result = executeQueryFieldOptions(controller, field, "an", 10);
-
-            Assertions.assertTrue(result.contains("banana"),
-                    "Expected the registered option matching the filter to "
-                            + "be returned, got: " + result);
-            Assertions.assertFalse(result.contains("Error"),
-                    "Tool must not error for a field that was registered "
-                            + "with valueOptions, got: " + result);
-        }
-
-        @Test
-        void queryFieldOptionsReportsUnknownFieldId() {
-            // When the LLM sends a field id the controller doesn't recognize
-            // (hallucinated, stale, or for a field that was registered with
-            // describe()/ignore() but not valueOptions), the tool must
-            // surface a specific 'unknown field id' error including the id
-            // itself so the LLM can correlate parallel tool calls and
-            // recover.
-            var registered = new TestField();
-            var unregistered = new TestField();
-            var controller = new FormAIController(
-                    new Div(registered, unregistered));
-            controller.valueOptions(registered, List.of("apple"),
-                    Function.identity());
-            controller.onRequestStart();
-
-            var resultUnknownId = executeQueryFieldOptions(controller,
-                    json("{\"field\":\"not-a-real-id\",\"filter\":\"\"}"));
-
-            Assertions.assertTrue(resultUnknownId.contains("Unknown field id"),
-                    "Hallucinated field id should produce the unknown-id "
-                            + "error message, got: " + resultUnknownId);
-            Assertions.assertTrue(resultUnknownId.contains("not-a-real-id"),
-                    "Error message should echo the offending id so the LLM "
-                            + "can correlate, got: " + resultUnknownId);
-
-            var unregisteredId = (String) ComponentUtil.getData(unregistered,
-                    FormAIController.FIELD_ID_KEY);
-            var resultFieldWithoutOptions = executeQueryFieldOptions(controller,
-                    unregistered, "", 10);
-
-            Assertions.assertTrue(
-                    resultFieldWithoutOptions.contains("Unknown field id"),
-                    "Field that was never registered with valueOptions "
-                            + "should produce the same unknown-id error, "
-                            + "got: " + resultFieldWithoutOptions);
-            Assertions.assertTrue(
-                    resultFieldWithoutOptions.contains(unregisteredId),
-                    "Error message should echo the offending id, got: "
-                            + resultFieldWithoutOptions);
-        }
-
-        @Test
-        void queryFieldOptionsRecoversToDefaultLimitForNonPositiveLimit() {
-            // A misbehaving LLM may send limit=0 (or negative). The tool
-            // must recover to the default rather than forward the bogus
-            // value to the query callback. Drive end-to-end via the same
-            // path the LLM uses.
-            var field = new TestField();
-            var capturedLimit = new AtomicInteger();
-            var controller = new FormAIController(new Div(field));
-            controller.valueOptions(field, (filter, limit) -> {
-                capturedLimit.set(limit);
-                return List.of();
-            }, Function.identity());
-            controller.onRequestStart();
-
-            executeQueryFieldOptions(controller, field, "", 0);
-
-            Assertions.assertEquals(50, capturedLimit.get(),
-                    "Non-positive limit must be coerced to the default "
-                            + "before reaching the query callback, got: "
-                            + capturedLimit.get());
-        }
-
-        @Test
-        void queryFieldOptionsReturnsErrorForMissingFieldArgument() {
-            var controller = new FormAIController(new Div(new TestField()));
-
-            var result = executeQueryFieldOptions(controller,
-                    json("{\"filter\":\"\"}"));
-
-            Assertions.assertTrue(result.startsWith("Error"));
-            Assertions.assertTrue(result.contains("field"));
-        }
-
-        @Test
-        void queryFieldOptionsReturnsErrorForNullArguments() {
-            var controller = new FormAIController(new Div(new TestField()));
-
-            var result = executeQueryFieldOptions(controller, null);
-
-            Assertions.assertTrue(result.startsWith("Error"));
-        }
-
-        @Test
-        void queryFieldOptionsDefaultsFilterToEmptyAndLimitToFifty() {
-            var field = new TestField();
-            var capturedFilter = new AtomicReference<String>();
-            var capturedLimit = new AtomicInteger();
-            var controller = new FormAIController(new Div(field));
-            controller.valueOptions(field, (filter, limit) -> {
-                capturedFilter.set(filter);
-                capturedLimit.set(limit);
-                return List.of();
-            }, Function.identity());
-            controller.onRequestStart();
-
-            var fieldId = (String) ComponentUtil.getData(field,
-                    FormAIController.FIELD_ID_KEY);
-            executeQueryFieldOptions(controller,
-                    json("{\"field\":\"" + fieldId + "\"}"));
-
-            Assertions.assertEquals("", capturedFilter.get());
-            Assertions.assertEquals(50, capturedLimit.get());
-        }
-
-        @Test
-        void queryFieldOptionsForwardsFilterAndLimitToTheRegisteredQuery() {
-            var field = new TestField();
-            var capturedFilter = new AtomicReference<String>();
-            var capturedLimit = new AtomicInteger();
-            var controller = new FormAIController(new Div(field));
-            controller.valueOptions(field, (filter, limit) -> {
-                capturedFilter.set(filter);
-                capturedLimit.set(limit);
-                return List.of();
-            }, Function.identity());
-            controller.onRequestStart();
-
-            executeQueryFieldOptions(controller, field, "acme", 7);
-
-            Assertions.assertEquals("acme", capturedFilter.get());
-            Assertions.assertEquals(7, capturedLimit.get());
-        }
-
-        @Test
-        void queryFieldOptionsEmitsOneLinePerLabel() {
-            var field = new TestField();
-            var controller = new FormAIController(new Div(field));
-            controller.valueOptions(field,
-                    List.of("Apollo #P-1", "Polaris #P-2"),
-                    Function.identity());
-            controller.onRequestStart();
-
-            var result = executeQueryFieldOptions(controller, field, "", 50);
-
-            Assertions.assertEquals("Apollo #P-1\nPolaris #P-2\n", result);
-        }
-
-        @Test
-        void queryFieldOptionsClampsLimitToTwoHundred() {
-            var field = new TestField();
-            var capturedLimit = new AtomicInteger();
-            var controller = new FormAIController(new Div(field));
-            controller.valueOptions(field, (filter, limit) -> {
-                capturedLimit.set(limit);
-                // Return more items than the cap so truncation kicks in.
-                var items = new ArrayList<String>(201);
-                for (var i = 0; i < 201; i++) {
-                    items.add("item-" + i);
-                }
-                return items;
-            }, Function.identity());
-            controller.onRequestStart();
-
-            var result = executeQueryFieldOptions(controller, field, "", 9999);
-
-            Assertions.assertEquals(200, capturedLimit.get());
-            Assertions.assertTrue(result.contains("(truncated to 200 items)"),
-                    "Result should signal truncation, got: " + result);
-            var dataLines = Arrays.stream(result.split("\n"))
-                    .filter(s -> !s.isEmpty()).filter(s -> !s.startsWith("("))
-                    .count();
-            Assertions.assertEquals(200, dataLines,
-                    "Output must not contain more data lines than the "
-                            + "clamped limit even when the callback returns "
-                            + "more items, got: " + dataLines);
-        }
-
-        @Test
-        void queryFieldOptionsDoesNotClaimTruncationWhenResultsFitUnderLimit() {
-            // When the LLM requests a limit above the server cap, the cap
-            // kicks in — but if the callback returns far fewer items than
-            // the cap, the result was not actually truncated. The
-            // "(truncated to ... items)" message should only appear when
-            // items were dropped.
-            var field = new TestField();
-            var controller = new FormAIController(new Div(field));
-            controller.valueOptions(field,
-                    (filter, limit) -> List.of("only-one"),
-                    Function.identity());
-            controller.onRequestStart();
-
-            var result = executeQueryFieldOptions(controller, field, "", 9999);
-
-            Assertions.assertFalse(result.contains("truncated"),
-                    "Result must not claim truncation when fewer items "
-                            + "than the cap were returned, got: " + result);
-        }
-
-        @Test
-        void queryFieldOptionsSignalsEmptyResultExplicitly() {
-            // An empty body is indistinguishable from a broken tool to the
-            // LLM. When the query returns zero items, the result should
-            // carry an explicit signal rather than just "".
-            var field = new TestField();
-            var controller = new FormAIController(new Div(field));
-            controller.valueOptions(field, (filter, limit) -> List.of(),
-                    Function.identity());
-            controller.onRequestStart();
-
-            var result = executeQueryFieldOptions(controller, field, "zzz", 10);
-
-            Assertions.assertFalse(result.isEmpty(),
-                    "Empty match set must produce an explicit signal to "
-                            + "the LLM, not an empty string");
-        }
-
-        @Test
-        void queryFieldOptionsEscapesNewlinesInLabels() {
-            // Labels are emitted one-per-line. A label containing '\n'
-            // would silently corrupt the format, leaving the LLM unable to
-            // recover the original options. The tool escapes '\n' (and the
-            // escape char itself) in labels so a naive split by '\n'
-            // yields one entry per original label.
-            var field = new TestField();
-            var controller = new FormAIController(new Div(field));
-            controller.valueOptions(field,
-                    (filter, limit) -> List.of("first\nsecond", "third"),
-                    Function.identity());
-            controller.onRequestStart();
-
-            var result = executeQueryFieldOptions(controller, field, "", 10);
-
-            var lines = java.util.Arrays.stream(result.split("\n"))
-                    .filter(s -> !s.isEmpty()).toList();
-            Assertions.assertEquals(List.of("first\\nsecond", "third"), lines,
-                    "Labels containing newlines must be escaped so the "
-                            + "output format stays parseable, got: " + result);
-        }
-
-        @Test
-        void queryFieldOptionsDoesNotLeakRawExceptionContent() {
-            // Exception messages from a user-supplied query callback can
-            // contain sensitive data (JDBC URLs, file paths, upstream API
-            // bodies, tokens). The tool must not echo ex.getMessage() (or
-            // any uncontrolled exception content) into the response handed
-            // to the LLM.
-            var sentinel = "jdbc:postgresql://prod-db.internal:5432/secrets "
-                    + "TOKEN=abc123";
-            var field = new TestField();
-            var controller = new FormAIController(new Div(field));
-            controller.valueOptions(field, (filter, limit) -> {
-                throw new IllegalStateException(sentinel);
-            }, Function.identity());
-            controller.onRequestStart();
-
-            var result = executeQueryFieldOptions(controller, field, "", 10);
-
-            Assertions.assertTrue(result.startsWith("Error"),
-                    "Failures should surface as an error to the LLM, got: "
-                            + result);
-            Assertions.assertFalse(result.contains(sentinel),
-                    "Raw exception message must not be forwarded to the "
-                            + "LLM verbatim — it can leak internal "
-                            + "details. Got: " + result);
-            Assertions.assertFalse(
-                    result.contains("jdbc:") || result.contains("TOKEN="),
-                    "Fragments of the exception message must not leak "
-                            + "either, got: " + result);
-        }
-
-        @Test
-        void queryFieldOptionsSchemaIsStatic() {
-            // The parameters schema is built once and does not enumerate
-            // field ids — clients can cache it across requests.
-            var controller = new FormAIController(new Div(new TestField()));
-            var first = findTool(controller.getTools(), "query_field_options")
-                    .getParametersSchema();
-            var second = findTool(controller.getTools(), "query_field_options")
-                    .getParametersSchema();
-            Assertions.assertEquals(first, second);
-
-            var schema = json(first);
-            Assertions.assertTrue(
-                    schema.path("properties").path("field").path("enum")
-                            .isMissingNode(),
-                    "Static schema should not encode field ids as an enum");
-        }
-
-    }
-
-    @Nested
-    class FormState {
-
-        @Test
-        void getFormStateReturnsAllVisibleFieldsInDocumentOrder() {
-            var a = new TestField();
-            var b = new DoubleField();
-            var c = new BoolField();
-            var nested = new Div(b, c);
-            var form = new Div(a, nested);
-            var controller = new FormAIController(form);
-
-            var fields = formStateFields(controller);
-
-            Assertions.assertEquals(3, fields.size());
-            Assertions.assertEquals(idOf(a),
-                    fields.get(0).get("id").asString());
-            Assertions.assertEquals(idOf(b),
-                    fields.get(1).get("id").asString());
-            Assertions.assertEquals(idOf(c),
-                    fields.get(2).get("id").asString());
-        }
-
-        @Test
-        void getFormStateOmitsIgnoredFields() {
-            var visible = new TestField();
-            var hidden = new TestField();
-            var controller = new FormAIController(new Div(visible, hidden));
-            controller.ignore(hidden);
-
-            var fields = formStateFields(controller);
-
-            Assertions.assertEquals(1, fields.size());
-            Assertions.assertEquals(idOf(visible),
-                    fields.get(0).get("id").asString());
-        }
-
-        @Test
-        void getFormStateReturnsEmptyFieldsArrayForEmptyForm() {
-            var controller = new FormAIController(new Div());
-
-            var fields = formStateFields(controller);
-
-            Assertions.assertEquals(0, fields.size());
-        }
-
-        @Test
-        void getFormStateRendersBigIntegerWithoutOverflow() {
-            var field = new BigIntField();
-            var huge = new java.math.BigInteger("99999999999999999999");
-            field.setValue(huge);
-            var controller = new FormAIController(new Div(field));
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertEquals(huge, f.get("value").bigIntegerValue());
-        }
-
-        @Test
-        void getFormStateAutoIgnoresPasswordFieldEvenWithHints() {
-            var visible = new TestField();
-            var password = new PasswordField();
-            password.setValue("secret");
-            var controller = new FormAIController(new Div(visible, password))
-                    .describe(password, "Account password")
-                    .valueOptions(password, List.of("hunter2"));
-
-            var fields = formStateFields(controller);
-
-            Assertions.assertEquals(1, fields.size());
-            Assertions.assertEquals(idOf(visible),
-                    fields.get(0).get("id").asString());
-        }
-
-        @Test
-        void getFormStateAutoIgnoresPasswordField() {
-            // PasswordField is treated as UNSUPPORTED so secret values do
-            // not flow into the LLM tool payload by default. The developer
-            // does not have to remember to call .ignore() for every form
-            // that happens to include one.
-            var visible = new TestField();
-            var password = new PasswordField();
-            var controller = new FormAIController(new Div(visible, password));
-
-            var fields = formStateFields(controller);
-
-            Assertions.assertEquals(1, fields.size(),
-                    "PasswordField must never appear in the form-state "
-                            + "result, got: " + fields);
-            Assertions.assertEquals(idOf(visible),
-                    fields.get(0).get("id").asString());
-        }
-
-        @Test
-        void getFormStateMarksEmailFieldWithEmailFormat() {
-            // EmailField shares String as its value type with TextField, so
-            // the format=email distinction has to come from a class-name
-            // exception in the classifier rather than from HasValue's
-            // generic parameter.
-            var field = new EmailField();
-            var controller = new FormAIController(new Div(field));
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertEquals("string", f.path("type").asString());
-            Assertions.assertEquals("email", f.path("format").asString(),
-                    "EmailField must carry format=email, got: " + f);
-        }
-
-        @Test
-        void getFormStateMergesLabelDescriptionAndHelperText() {
-            var field = new LabeledStringField();
-            field.setLabel("Merchant");
-            field.setHelperText("As shown on the receipt");
-            var controller = new FormAIController(new Div(field));
-            controller.describe(field, "The vendor name");
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertEquals(
-                    "Merchant | The vendor name | As shown on the receipt",
-                    f.get("description").asString(),
-                    "label, registered description, and helper text must be "
-                            + "joined in that order");
-        }
-
-        @Test
-        void getFormStateMergeDescriptionKeepsTrailingPeriodsVerbatim() {
-            // Real-world labels and helper texts often end in '.', and
-            // user-supplied descriptions almost always do. The merge
-            // separator must not collide with trailing punctuation: a
-            // '. ' separator would produce "Merchant.. The vendor..".
-            // Pin that each part appears intact and no '..' shows up.
-            var field = new LabeledStringField();
-            field.setLabel("Merchant.");
-            field.setHelperText("As shown on the receipt.");
-            var controller = new FormAIController(new Div(field));
-            controller.describe(field, "The vendor name.");
-
-            var f = formStateFields(controller).get(0);
-            var desc = f.get("description").asString();
-
-            Assertions.assertFalse(desc.contains(".."),
-                    "Merge separator must not collide with trailing "
-                            + "periods, got: " + desc);
-            Assertions.assertTrue(desc.contains("Merchant."));
-            Assertions.assertTrue(desc.contains("The vendor name."));
-            Assertions.assertTrue(desc.contains("As shown on the receipt."));
-        }
-
-        @Test
-        void getFormStateOmitsBlankDescriptionEntirely() {
-            var field = new TestField();
-            var controller = new FormAIController(new Div(field));
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertFalse(f.has("description"),
-                    "Empty merged description must be omitted, got: " + f);
-        }
-
-        @Test
-        void getFormStateMapsStringValueTypeToTypeString() {
-            assertTypeOnly(typeNodeFor(new TestField()), "string");
-        }
-
-        @Test
-        void getFormStateMapsDoubleValueTypeToTypeNumber() {
-            assertTypeOnly(typeNodeFor(new DoubleField()), "number");
-        }
-
-        @Test
-        void getFormStateMapsIntegerValueTypeToTypeInteger() {
-            assertTypeOnly(typeNodeFor(new IntField()), "integer");
-        }
-
-        @Test
-        void getFormStateClassifiesDirectHasValueImplementation() {
-            // All other tests reach HasValue<?, V> through AbstractField →
-            // HasValueAndElement → HasValue, exercising type-variable
-            // propagation. This test covers the simpler case: a Component
-            // that declares its V argument directly on HasValue, with no
-            // intermediate generic chain.
-            assertTypeOnly(typeNodeFor(new DirectIntegerField()), "integer");
-        }
-
-        @Test
-        void getFormStateMapsBigDecimalValueTypeToStringWithPattern() {
-            assertTypeAndPattern(typeNodeFor(new BigDecField()), "string",
-                    "^-?\\d+(\\.\\d+)?$");
-        }
-
-        @Test
-        void getFormStateMapsBooleanValueTypeToTypeBoolean() {
-            assertTypeOnly(typeNodeFor(new BoolField()), "boolean");
-        }
-
-        @Test
-        void getFormStateMapsLocalDateValueTypeToStringWithDateFormat() {
-            assertTypeAndFormat(typeNodeFor(new DateField()), "string", "date");
-        }
-
-        @Test
-        void getFormStateMapsLocalDateTimeValueTypeToStringWithDateTimeFormat() {
-            assertTypeAndFormat(typeNodeFor(new DateTimeField()), "string",
-                    "date-time");
-        }
-
-        @Test
-        void getFormStateMapsLocalTimeValueTypeToStringWithTimeFormat() {
-            assertTypeAndFormat(typeNodeFor(new TimeField()), "string", "time");
-        }
-
-        @Test
-        void getFormStateMapsLongValueTypeToTypeInteger() {
-            // Integer is the obvious INTEGER case; Long is the case that
-            // breaks when the classifier narrows its INTEGER detection to
-            // Integer + BigInteger only. Pin Long classification separately
-            // so a regression there cannot hide behind the IntField test.
-            assertTypeOnly(typeNodeFor(new LongField()), "integer");
-        }
-
-        @Test
-        void getFormStateMapsShortValueTypeToTypeInteger() {
-            assertTypeOnly(typeNodeFor(new ShortField()), "integer");
-        }
-
-        @Test
-        void getFormStateMapsByteValueTypeToTypeInteger() {
-            assertTypeOnly(typeNodeFor(new ByteField()), "integer");
-        }
-
-        @Test
-        void getFormStateMapsFloatValueTypeToTypeNumber() {
-            // Float-valued HasValue fields must classify the same as Double
-            // (NUMBER) — both round-trip through JSON's number type. A
-            // classifier that recognises only Double would silently demote
-            // Float fields to STRING.
-            assertTypeOnly(typeNodeFor(new FloatField()), "number");
-        }
-
-        @Test
-        void getFormStateResolvesValueTypeThroughNonGenericIntermediateInterface() {
-            // The reflective walk that resolves HasValue's V argument has
-            // two branches: one for parameterized intermediate interfaces,
-            // one for raw classes that themselves extend a parameterized
-            // HasValue (e.g. `interface MyField extends HasValue<E, Integer>
-            // {}`). The raw-class branch must propagate the found type out
-            // of the recursive call; if it discards the result, the field
-            // is silently misclassified as STRING.
-            assertTypeOnly(
-                    typeNodeFor(new IntegerViaNonGenericInterfaceField()),
-                    "integer");
-        }
-
-        @Test
-        void getFormStateClassifiesHasLazyDataViewFieldAsSingleSelect() {
-            // Real Vaadin selection components implement the data-view
-            // marker interfaces (HasLazyDataView / HasListDataView /
-            // HasDataView) but not HasItems. The classifier must accept
-            // any of these markers as a selection signal — otherwise the
-            // field is treated as STRING and applyType skips the path
-            // that enumerates the ListDataProvider's items as `enum`.
-            // The visible-from-tests difference is precisely that enum
-            // block.
-            var field = new LazyDataViewField();
-            field.supplyItems("alpha", "beta");
-            var controller = new FormAIController(new Div(field));
-
-            var f = formStateFields(controller).get(0);
-
-            var values = new ArrayList<String>();
-            f.path("enum").forEach(n -> values.add(n.asString()));
-            Assertions.assertEquals(List.of("alpha", "beta"), values,
-                    "Field with HasLazyDataView marker must enumerate its "
-                            + "ListDataProvider items as `enum`; an empty "
-                            + "or missing enum means the field fell through "
-                            + "to the non-selection branch. Got: " + f);
-        }
-
-        @Test
-        void getFormStateClassifiesHasListDataViewFieldAsSingleSelect() {
-            var field = new ListDataViewField();
-            field.supplyItems("apple", "banana");
-            var controller = new FormAIController(new Div(field));
-
-            var f = formStateFields(controller).get(0);
-
-            var values = new ArrayList<String>();
-            f.path("enum").forEach(n -> values.add(n.asString()));
-            Assertions.assertEquals(List.of("apple", "banana"), values,
-                    "Field with HasListDataView marker must enumerate its "
-                            + "ListDataProvider items as `enum`. Got: " + f);
-        }
-
-        @Test
-        void getFormStateClassifiesHasDataViewFieldAsSingleSelect() {
-            var field = new DataViewField();
-            field.supplyItems("x", "y");
-            var controller = new FormAIController(new Div(field));
-
-            var f = formStateFields(controller).get(0);
-
-            var values = new ArrayList<String>();
-            f.path("enum").forEach(n -> values.add(n.asString()));
-            Assertions.assertEquals(List.of("x", "y"), values,
-                    "Field with HasDataView marker must enumerate its "
-                            + "ListDataProvider items as `enum`. Got: " + f);
-        }
-
-        @Test
-        void getFormStateMultiSelectDoesNotSetTypeAtNodeLevel() {
-            // MultiSelect's schema lives entirely in the items block:
-            // `{ array: true, items: { type: "string", ... } }`. The node
-            // itself must not also carry `type` — a duplicate type at the
-            // node level would conflict with the array shape and confuse
-            // strict schema consumers.
-            var f = typeNodeFor(new MultiSelectField<String>());
-
-            Assertions.assertTrue(f.path("array").asBoolean());
-            Assertions.assertTrue(f.path("type").isMissingNode(),
-                    "Multi-select must not duplicate type at the node level "
-                            + "(it belongs inside items), got: " + f);
-        }
-
-        @Test
-        void getFormStateMultiSelectWithValueOptionsKeepsNodeShape() {
-            // Combining multi-select with valueOptions must still leave the
-            // node-level shape clean: array=true, no node-level type, and
-            // queryable/enum sit inside the items block.
-            var multi = new MultiSelectField<String>();
-            var controller = new FormAIController(new Div(multi));
-            controller.valueOptions(multi, (filter, limit) -> List.of("a", "b"),
-                    Set::of);
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertTrue(f.path("array").asBoolean());
-            Assertions.assertTrue(f.path("type").isMissingNode(),
-                    "Multi-select with valueOptions must not set node-level "
-                            + "type, got: " + f);
-            Assertions.assertTrue(f.path("queryable").isMissingNode(),
-                    "queryable must live inside items, not on the node, "
-                            + "got: " + f);
-            Assertions.assertTrue(f.path("items").path("queryable").asBoolean(),
-                    "Queryable signal must sit on items, got: " + f);
-        }
-
-        @Test
-        void getFormStateMultiSelectWithValueOptionsRendersValueAsArray() {
-            // Multi-select value is a Set; serialising it as a single
-            // string (the path taken when valueOptions on non-selection
-            // fields rewrites to string) would corrupt the payload. Pin
-            // that multi-select wins over the valueOptions value-rewrite.
-            var multi = new MultiSelectField<String>();
-            multi.setItems("a", "b", "c");
-            multi.setValue(Set.of("a", "c"));
-            var controller = new FormAIController(new Div(multi));
-            controller.valueOptions(multi,
-                    (filter, limit) -> List.of("a", "b", "c"), Set::of);
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertTrue(f.path("value").isArray(),
-                    "Multi-select with valueOptions must still render its "
-                            + "value as a JSON array, got: " + f.path("value"));
-            var rendered = new ArrayList<String>();
-            f.path("value").forEach(n -> rendered.add(n.asString()));
-            Assertions.assertEquals(2, rendered.size());
-            Assertions.assertTrue(rendered.containsAll(List.of("a", "c")),
-                    "Array must carry the selected labels, got: " + rendered);
-        }
-
-        @Test
-        void getFormStateSingleSelectWithValueOptionsRendersValueAsLabel() {
-            // The mirror case for the multi-select test above: the
-            // selection-aware path must keep using the field's own label
-            // renderer for the current value, not the generic valueOptions
-            // value-rewrite branch.
-            var combo = new SingleSelectField<Project>();
-            combo.setItems(new Project("P-1", "Alpha"),
-                    new Project("P-2", "Beta"));
-            combo.setItemLabelGenerator(p -> p.code() + " " + p.name());
-            combo.setValue(new Project("P-2", "Beta"));
-            var controller = new FormAIController(new Div(combo));
-            controller.valueOptions(combo,
-                    (filter, limit) -> List.of("P-1 Alpha", "P-2 Beta"),
-                    label -> null);
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertEquals("P-2 Beta", f.path("value").asString(),
-                    "Single-select value must come from the field's label "
-                            + "generator, not Object#toString(), got: "
-                            + f.path("value"));
-        }
-
-        @Test
-        void getFormStateDescribedFieldHasNoQueryableFlag() {
-            // describe() registers a hint without a valueOptions callback.
-            // The schema must distinguish "has a hint entry" from "has a
-            // query callback": only the latter is queryable. Otherwise the
-            // LLM would call query_field_options against a field that has
-            // no registered query and get a generic error back.
-            // Use a selection-typed field so applySelectionOptions runs;
-            // a plain STRING field skips that branch entirely and would
-            // not exercise this regression.
-            var combo = new SingleSelectField<String>();
-            var controller = new FormAIController(new Div(combo));
-            controller.describe(combo, "Some descriptive text");
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertTrue(f.path("queryable").isMissingNode(),
-                    "A field with describe() but no valueOptions must not "
-                            + "carry queryable=true, got: " + f);
-        }
-
-        @Test
-        void getFormStateQueryableFieldDoesNotMixListDataProviderEnum() {
-            // When valueOptions(BiFunction) is registered on a selection
-            // component whose backing ListDataProvider also has items,
-            // queryable=true must short-circuit before the data-provider
-            // items would be enumerated as enum. Both signals in the same
-            // payload would tell the LLM "either query me, or pick from
-            // this list" — the registration says only one is authoritative.
-            var combo = new SingleSelectField<String>();
-            combo.setItems("apple", "banana", "cherry");
-            var controller = new FormAIController(new Div(combo));
-            controller.valueOptions(combo,
-                    (filter, limit) -> List.of("apple", "banana"),
-                    Function.identity());
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertTrue(f.path("queryable").asBoolean());
-            Assertions.assertTrue(f.path("enum").isMissingNode(),
-                    "queryable signal must suppress the ListDataProvider "
-                            + "enum fallback, got: " + f);
-        }
-
-        @Test
-        void getFormStateRendersNonFiniteNumericValueAsExplicitJsonNull() {
-            // Sister test to getFormStateRendersNumericInfinityAsFinite-
-            // JsonNumber: that one only checks the raw output lacks
-            // "Infinity"/"NaN" tokens, which is also satisfied by emitting
-            // no value key at all. Pin explicitly that the key is present
-            // and JSON null, so missing-key regressions don't slip through.
-            var field = new DoubleField();
-            field.setValue(Double.NaN);
-            var controller = new FormAIController(new Div(field));
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertTrue(f.has("value"),
-                    "value key must be present even for non-finite numerics, "
-                            + "got: " + f);
-            Assertions.assertTrue(f.path("value").isNull(),
-                    "Non-finite numeric must serialize as JSON null, got: "
-                            + f.path("value"));
-        }
-
-        @Test
-        void getFormStateRendersListDataProviderItemsViaLabelGenerator() {
-            // ListDataProvider items typically aren't strings — a
-            // ComboBox<Project> backed by a list of Project records needs
-            // its ItemLabelGenerator to be invoked when the enum block is
-            // built. Without the generator, items render via toString(),
-            // which leaks Java internals (e.g. "Project[code=P-1, ...]") to
-            // the LLM.
-            var combo = new SingleSelectField<Project>();
-            combo.setItems(new Project("P-1", "Alpha"),
-                    new Project("P-2", "Beta"));
-            combo.setItemLabelGenerator(p -> p.code() + " " + p.name());
-            var controller = new FormAIController(new Div(combo));
-
-            var f = formStateFields(controller).get(0);
-
-            var values = new ArrayList<String>();
-            f.path("enum").forEach(n -> values.add(n.asString()));
-            Assertions.assertEquals(List.of("P-1 Alpha", "P-2 Beta"), values,
-                    "Enum entries must use the field's label generator "
-                            + "output, got: " + values);
-        }
-
-        @Test
-        void getFormStateMapsSingleSelectFieldToStringType() {
-            // A selection component without registered options or items
-            // surfaces as bare "type": "string" — no enum, no queryable.
-            var f = typeNodeFor(new SingleSelectField<String>());
-
-            Assertions.assertEquals("string", f.path("type").asString());
-            Assertions.assertTrue(f.path("enum").isMissingNode());
-            Assertions.assertTrue(f.path("queryable").isMissingNode());
-        }
-
-        @Test
-        void getFormStateMapsMultiSelectFieldToArrayWithItemsBlock() {
-            var f = typeNodeFor(new MultiSelectField<String>());
-
-            Assertions.assertTrue(f.path("array").asBoolean(),
-                    "MultiSelect must be encoded with array=true");
-            Assertions.assertEquals("string",
-                    f.path("items").path("type").asString());
-        }
-
-        @Test
-        void getFormStateEncodesEnumForFixedValueOptions() {
-            // Fixed valueOptions registered against a selection component
-            // surface in the JSON output as enum. The same registration
-            // against a non-selection field would not (LLM uses
-            // query_field_options instead).
-            var combo = new SingleSelectField<String>();
-            var controller = new FormAIController(new Div(combo));
-            controller.valueOptions(combo, List.of("EUR", "USD", "GBP"));
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertEquals("string", f.path("type").asString());
-            Assertions.assertTrue(f.path("queryable").isMissingNode(),
-                    "Fixed-options field must not carry queryable=true");
-            var values = new ArrayList<String>();
-            f.path("enum").forEach(n -> values.add(n.asString()));
-            Assertions.assertEquals(List.of("EUR", "USD", "GBP"), values);
-        }
-
-        @Test
-        void getFormStateEncodesQueryableForBiFunctionValueOptions() {
-            var combo = new SingleSelectField<String>();
-            var controller = new FormAIController(new Div(combo));
-            controller.valueOptions(combo,
-                    (filter, limit) -> List.of("Apollo", "Polaris"));
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertTrue(f.path("queryable").asBoolean(),
-                    "Queryable-options field must carry queryable=true, "
-                            + "got: " + f);
-            Assertions.assertTrue(f.path("enum").isMissingNode(),
-                    "Queryable-options field must not enumerate options "
-                            + "inline, got: " + f);
-        }
-
-        @Test
-        void getFormStateExposesEnumForFixedValueOptionsOnStringField() {
-            var field = new TestField();
-            var controller = new FormAIController(new Div(field));
-            controller.valueOptions(field, List.of("EUR", "USD", "GBP"));
-
-            var f = formStateFields(controller).get(0);
-
-            var values = new ArrayList<String>();
-            f.path("enum").forEach(n -> values.add(n.asString()));
-            Assertions.assertEquals(List.of("EUR", "USD", "GBP"), values);
-        }
-
-        @Test
-        void getFormStateExposesQueryableForBiFunctionValueOptionsOnStringField() {
-            var field = new TestField();
-            var controller = new FormAIController(new Div(field));
-            controller.valueOptions(field,
-                    (filter, limit) -> List.of("Apollo", "Polaris"));
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertTrue(f.path("queryable").asBoolean());
-        }
-
-        @Test
-        void getFormStateUsesFallbackWhenLabelGeneratorReturnsNull() {
-            var combo = new SingleSelectField<String>();
-            combo.setItems("only");
-            combo.setItemLabelGenerator(item -> null);
-            combo.setValue("only");
-            var controller = new FormAIController(new Div(combo));
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertFalse(f.path("value").isNull());
-        }
-
-        @Test
-        void getFormStateExposesListDataProviderItemsAsEnum() {
-            var combo = new SingleSelectField<String>();
-            combo.setItems("alpha", "beta", "gamma");
-            var controller = new FormAIController(new Div(combo));
-
-            var f = formStateFields(controller).get(0);
-
-            var values = new ArrayList<String>();
-            f.path("enum").forEach(n -> values.add(n.asString()));
-            Assertions.assertEquals(List.of("alpha", "beta", "gamma"), values,
-                    "ListDataProvider items must populate the enum when no "
-                            + "valueOptions hint is registered");
-        }
-
-        @Test
-        void getFormStateEnumIsCappedForLargeListDataProvider() {
-            var combo = new SingleSelectField<String>();
-            var items = new ArrayList<String>();
-            for (var i = 0; i < 250; i++) {
-                items.add("item-" + i);
-            }
-            combo.setItems(items);
-            var controller = new FormAIController(new Div(combo));
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertTrue(f.path("enum").size() <= 200,
-                    "ListDataProvider items must be capped at 200 entries "
-                            + "to match the query_field_options cap, got: "
-                            + f.path("enum").size());
-        }
-
-        @Test
-        void getFormStateBigDecimalValueMatchesPatternForScientificInput() {
-            var field = new BigDecField();
-            field.setValue(new java.math.BigDecimal("1E2"));
-            var controller = new FormAIController(new Div(field));
-
-            var f = formStateFields(controller).get(0);
-            var rendered = f.get("value").asString();
-
-            Assertions.assertTrue(rendered.matches("^-?\\d+(\\.\\d+)?$"),
-                    "Rendered BigDecimal value must satisfy the schema "
-                            + "pattern, got: " + rendered);
-        }
-
-        @Test
-        void getFormStateRendersNumericInfinityAsFiniteJsonNumber() {
-            var field = new DoubleField();
-            field.setValue(Double.POSITIVE_INFINITY);
-            var controller = new FormAIController(new Div(field));
-
-            var raw = findTool(controller.getTools(), "get_form_state")
-                    .execute(JacksonUtils.createObjectNode());
-
-            Assertions.assertFalse(
-                    raw.contains("Infinity") || raw.contains("NaN"),
-                    "Tool result must not embed non-standard JSON tokens "
-                            + "for non-finite doubles, got: " + raw);
-        }
-
-        @Test
-        void getFormStateBackendDataProviderProducesNoEnum() {
-            var combo = new SingleSelectField<String>();
-            combo.setDataProvider(new CallbackDataProvider<String, String>(
-                    q -> List.of("a", "b").stream(), q -> 2));
-            var controller = new FormAIController(new Div(combo));
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertEquals("string", f.path("type").asString());
-            Assertions.assertTrue(f.path("enum").isMissingNode(),
-                    "Backend data provider must not contribute an enum");
-            Assertions.assertTrue(f.path("queryable").isMissingNode(),
-                    "Backend data provider alone must not flag queryable");
-        }
-
-        @Test
-        void getFormStateRendersCurrentValuesPerType() {
-            var text = new TestField();
-            text.setValue("Trattoria");
-            var integer = new IntField();
-            integer.setValue(42);
-            var number = new DoubleField();
-            number.setValue(58.4);
-            var bigDecimal = new BigDecField();
-            bigDecimal.setValue(new BigDecimal("58.40"));
-            var date = new DateField();
-            date.setValue(LocalDate.of(2026, 5, 4));
-            var bool = new BoolField();
-            bool.setValue(true);
-            var combo = new SingleSelectField<String>();
-            combo.setItems("Meals", "Travel");
-            combo.setValue("Meals");
-            var controller = new FormAIController(new Div(text, integer, number,
-                    bigDecimal, date, bool, combo));
-
-            var fields = formStateFields(controller);
-
-            Assertions.assertEquals("Trattoria",
-                    fields.get(0).get("value").asString());
-            Assertions.assertEquals(42, fields.get(1).get("value").asInt());
-            Assertions.assertEquals(58.4, fields.get(2).get("value").asDouble(),
-                    1e-9);
-            Assertions.assertEquals("58.40",
-                    fields.get(3).get("value").asString());
-            Assertions.assertEquals("2026-05-04",
-                    fields.get(4).get("value").asString());
-            Assertions.assertTrue(fields.get(5).get("value").asBoolean());
-            Assertions.assertEquals("Meals",
-                    fields.get(6).get("value").asString());
-        }
-
-        @Test
-        void getFormStateContinuesPastFieldWhoseGetValueThrows() {
-            var good = new TestField();
-            good.setValue("ok");
-            var bad = new ThrowingField();
-            var trailing = new TestField();
-            trailing.setValue("trailing");
-            var controller = new FormAIController(new Div(good, bad, trailing));
-
-            var fields = formStateFields(controller);
-
-            Assertions.assertEquals(3, fields.size(),
-                    "Failing field must still produce an entry so the LLM "
-                            + "can see that the field exists");
-            Assertions.assertEquals("ok",
-                    fields.get(0).path("value").asString(),
-                    "Fields before a failing field must still be present");
-            Assertions.assertEquals("trailing",
-                    fields.get(2).path("value").asString(),
-                    "Fields after a failing field must still be present");
-            Assertions.assertEquals(idOf(bad),
-                    fields.get(1).path("id").asString(),
-                    "Failed field entry must carry its id so the LLM can "
-                            + "correlate, got: " + fields.get(1));
-            Assertions.assertFalse(fields.get(1).path("error").isMissingNode(),
-                    "Failed field entry must carry an error marker, got: "
-                            + fields.get(1));
-        }
-
-        @Test
-        void getFormStateRendersStringValueWhenValueOptionsOverrideType() {
-            // valueOptions on a non-String field rewrites the schema type to
-            // "string" + enum, but applyValue keeps following the field's
-            // original FormFieldType — so the value half of the payload
-            // disagrees with the schema half. A strict consumer validating
-            // type=string against a JSON integer would reject the result.
-            var field = new IntField();
-            field.setValue(2);
-            var controller = new FormAIController(new Div(field));
-            controller.valueOptions(field, List.of("1", "2", "3"),
-                    Integer::parseInt);
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertEquals("string", f.path("type").asString());
-            Assertions.assertTrue(f.path("value").isString(),
-                    "value must match the declared schema type, got: "
-                            + f.path("value"));
-        }
-
-        @Test
-        void getFormStateRendersNullForEmptyValues() {
-            var text = new TestField();
-            var date = new DateField();
-            var multi = new MultiSelectField<String>();
-            var controller = new FormAIController(new Div(text, date, multi));
-
-            var fields = formStateFields(controller);
-
-            Assertions.assertTrue(fields.get(0).get("value").isNull(),
-                    "Empty text value must serialize as JSON null");
-            Assertions.assertTrue(fields.get(1).get("value").isNull(),
-                    "Empty date value must serialize as JSON null");
-            Assertions.assertTrue(fields.get(2).get("value").isNull(),
-                    "Empty multi-select value must serialize as JSON null");
-        }
-
-        @Test
-        void getFormStateMultiSelectUsesItemsBlock() {
-            var multi = new MultiSelectField<String>();
-            multi.setItems("a", "b", "c");
-            multi.setValue(Set.of("a", "c"));
-            var controller = new FormAIController(new Div(multi));
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertTrue(f.path("array").asBoolean());
-            Assertions.assertEquals("string",
-                    f.path("items").path("type").asString());
-            var enumValues = new ArrayList<String>();
-            f.path("items").path("enum")
-                    .forEach(n -> enumValues.add(n.asString()));
-            Assertions.assertTrue(
-                    enumValues.containsAll(List.of("a", "b", "c")),
-                    "Items block must enumerate the data-provider labels, "
-                            + "got: " + enumValues);
-            var rendered = new ArrayList<String>();
-            f.path("value").forEach(n -> rendered.add(n.asString()));
-            Assertions.assertTrue(
-                    rendered.containsAll(List.of("a", "c"))
-                            && rendered.size() == 2,
-                    "Multi-select value must render as a JSON array of label "
-                            + "strings, got: " + rendered);
-        }
-
-        @Test
-        void getFormStateMultiSelectQueryableUsesItemsQueryable() {
-            var multi = new MultiSelectField<String>();
-            var controller = new FormAIController(new Div(multi));
-            controller.valueOptions(multi, (filter, limit) -> List.of("x"),
-                    Set::of);
-
-            var f = formStateFields(controller).get(0);
-
-            Assertions.assertTrue(f.path("array").asBoolean());
-            Assertions.assertTrue(f.path("items").path("queryable").asBoolean(),
-                    "Queryable multi-select must carry queryable=true on the "
-                            + "items block, got: " + f);
-            Assertions.assertTrue(f.path("items").path("enum").isMissingNode(),
-                    "Queryable multi-select must not enumerate options "
-                            + "inline, got: " + f);
-        }
-
-        @Test
-        void getFormStateSchemaIsStaticAndEmpty() {
-            var controller = new FormAIController(new Div(new TestField()));
-
-            var schema = findTool(controller.getTools(), "get_form_state")
-                    .getParametersSchema();
-            var node = json(schema);
-
-            Assertions.assertEquals("object", node.path("type").asString());
-            Assertions.assertTrue(node.path("properties").isObject());
-            Assertions.assertEquals(0, node.path("properties").size(),
-                    "get_form_state must take no parameters");
-        }
-
-        @Test
-        void getFormStateReflectsLiveLabelChanges() {
-            var field = new LabeledStringField();
-            field.setLabel("Old");
-            var controller = new FormAIController(new Div(field));
-
-            Assertions.assertEquals("Old", formStateFields(controller).get(0)
-                    .get("description").asString());
-
-            field.setLabel("New");
-
-            Assertions.assertEquals("New",
-                    formStateFields(controller).get(0).get("description")
-                            .asString(),
-                    "description must be re-read from the live field on each "
-                            + "call");
-        }
-
-        @Test
-        void getFormStateAssignsStableIdAcrossCalls() {
-            var field = new TestField();
-            var controller = new FormAIController(new Div(field));
-
-            var firstId = formStateFields(controller).get(0).get("id")
-                    .asString();
-            var secondId = formStateFields(controller).get(0).get("id")
-                    .asString();
-
-            Assertions.assertEquals(firstId, secondId);
-            Assertions.assertEquals(firstId, idOf(field),
-                    "id in get_form_state must match the id stored on the "
-                            + "component");
-        }
-
-        @Test
-        void getFormStateProducesCompletePayloadForRealisticForm() {
-            // End-to-end smoke test that mirrors the RFC's expense-form
-            // example. Catches regressions in how the per-field slices
-            // interact when N > 1 — schema/value/description leaks across
-            // fields, ordering bugs that only surface with multiple
-            // entries, hint registrations affecting the wrong id.
-            var merchant = new LabeledStringField();
-            merchant.setLabel("Merchant");
-            merchant.setHelperText("As shown on the receipt");
-            merchant.setValue("Trattoria Toscana");
-
-            var amount = new DoubleField();
-            var currency = new SingleSelectField<String>();
-            var date = new DateField();
-            var category = new SingleSelectField<String>();
-            var notes = new TestField();
-
-            var form = new Div(merchant, amount, currency, date, category,
-                    notes);
-            var controller = new FormAIController(form)
-                    .describe(merchant, "The vendor name")
-                    .valueOptions(currency, List.of("EUR", "USD", "GBP"))
-                    .valueOptions(category, List.of("Travel", "Meals",
-                            "Software", "Office", "Other"));
-
-            // Execute first so the controller walks the form and assigns
-            // ids to fields that had no hints registered.
-            var actual = json(findTool(controller.getTools(), "get_form_state")
-                    .execute(JacksonUtils.createObjectNode()));
-
-            var expected = json(
-                    """
-                            {
-                              "fields": [
-                                {
-                                  "id": "<merchant>",
-                                  "description": "Merchant | The vendor name | As shown on the receipt",
-                                  "type": "string",
-                                  "value": "Trattoria Toscana"
-                                },
-                                {
-                                  "id": "<amount>",
-                                  "type": "number",
-                                  "value": null
-                                },
-                                {
-                                  "id": "<currency>",
-                                  "type": "string",
-                                  "enum": ["EUR", "USD", "GBP"],
-                                  "value": null
-                                },
-                                {
-                                  "id": "<date>",
-                                  "type": "string",
-                                  "format": "date",
-                                  "value": null
-                                },
-                                {
-                                  "id": "<category>",
-                                  "type": "string",
-                                  "enum": ["Travel", "Meals", "Software", "Office", "Other"],
-                                  "value": null
-                                },
-                                {
-                                  "id": "<notes>",
-                                  "type": "string",
-                                  "value": null
-                                }
-                              ]
-                            }"""
-                            .replace("<merchant>", idOf(merchant))
-                            .replace("<amount>", idOf(amount))
-                            .replace("<currency>", idOf(currency))
-                            .replace("<date>", idOf(date))
-                            .replace("<category>", idOf(category))
-                            .replace("<notes>", idOf(notes)));
-
-            Assertions.assertEquals(expected, actual);
-        }
-
-        @Test
-        void getToolsAlwaysIncludesFormState() {
-            // Unlike query_field_options, the form-state tool is always
-            // present — there is no per-field configuration that gates it.
-            var controllerNoHints = new FormAIController(new Div());
-            Assertions.assertTrue(
-                    controllerNoHints.getTools().stream().anyMatch(
-                            t -> t.getName().equals("get_form_state")),
-                    "get_form_state must be exposed even when the form has "
-                            + "no fields");
-
-            var field = new TestField();
-            var controllerWithIgnored = new FormAIController(new Div(field));
-            controllerWithIgnored.ignore(field);
-            Assertions.assertTrue(
-                    controllerWithIgnored.getTools().stream().anyMatch(
-                            t -> t.getName().equals("get_form_state")),
-                    "get_form_state must be exposed even when every field "
-                            + "is ignored");
-        }
-
-        // ---- helpers scoped to FormState ----
-
-        private JsonNode typeNodeFor(HasValue<?, ?> field) {
-            var controller = new FormAIController(new Div((Component) field));
-            return formStateFields(controller).get(0);
-        }
-
-        private List<JsonNode> formStateFields(FormAIController controller) {
-            var result = findTool(controller.getTools(), "get_form_state")
-                    .execute(JacksonUtils.createObjectNode());
-            var root = json(result);
-            var out = new ArrayList<JsonNode>();
-            root.path("fields").forEach(out::add);
-            return out;
-        }
-
-        private String idOf(HasValue<?, ?> field) {
-            return (String) ComponentUtil.getData((Component) field,
-                    FormAIController.FIELD_ID_KEY);
-        }
-
-        private void assertTypeOnly(JsonNode field, String expectedType) {
-            Assertions.assertEquals(expectedType, field.path("type").asString(),
-                    "Expected type=" + expectedType + " for " + field);
-            Assertions.assertTrue(field.path("format").isMissingNode(),
-                    "Expected no format for " + field);
-            Assertions.assertTrue(field.path("pattern").isMissingNode(),
-                    "Expected no pattern for " + field);
-        }
-
-        private void assertTypeAndFormat(JsonNode field, String expectedType,
-                String expectedFormat) {
-            Assertions.assertEquals(expectedType,
-                    field.path("type").asString());
-            Assertions.assertEquals(expectedFormat,
-                    field.path("format").asString());
-        }
-
-        private void assertTypeAndPattern(JsonNode field, String expectedType,
-                String expectedPattern) {
-            Assertions.assertEquals(expectedType,
-                    field.path("type").asString());
-            Assertions.assertEquals(expectedPattern,
-                    field.path("pattern").asString());
-        }
-    }
-
-    // --- Helpers ---
-
-    private static String executeQueryFieldOptions(FormAIController controller,
-            HasValue<?, ?> field, String filter, int limit) {
-        var fieldId = (String) ComponentUtil.getData((Component) field,
-                FormAIController.FIELD_ID_KEY);
-        return executeQueryFieldOptions(controller,
-                json("{\"field\":\"" + fieldId + "\",\"filter\":\"" + filter
-                        + "\",\"limit\":" + limit + "}"));
-    }
-
-    private static String executeQueryFieldOptions(FormAIController controller,
-            JsonNode arguments) {
-        return findTool(controller.getTools(), "query_field_options")
-                .execute(arguments);
-    }
-
-    private static LLMProvider.ToolSpec findTool(
-            List<LLMProvider.ToolSpec> tools, String name) {
-        return tools.stream().filter(t -> t.getName().equals(name)).findFirst()
-                .orElseThrow();
-    }
-
-    private static JsonNode json(String text) {
-        return JacksonUtils.readTree(text);
     }
 }
